// models/Hotel.ts
export interface Hotel {
    id: string;
    hotelName: string;
    hotelCity: string;
    hotelAddress: string;
    hotelState: string;
    hotelPincode: string;
    hotelEmailId: string;
    hotelContactNumber: number;
    hotelStarRating: number;
    hotelImageUrl: string;
    createdAt: string;
    updatedAt: string;
    hotelImagesList:[],
    hotelSlug:''
}